requirements.txt
tensorflow>=2.0.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
numpy>=1.19.0
pyyaml>=5.4.1
wandb>=0.12.0